+++ adminer/drivers/mssql.inc.php
@@ -281,7 +281,8 @@ if (isset($_GET["mssql"])) {
 				function attach(array $server, string $username, string $password): string {
 					$port = $server["port"];
 					$socket = $server["socket"];
-					return $this->dsn("dblib:charset=utf8;host=$server[host]" . ($port != "" ? ";port=$port" : ($socket != "" ? ";unix_socket=$socket" : "")), $username, $password);
+					$options = array(1002 => true); // 1002 - PDO::DBLIB_ATTR_STRINGIFY_UNIQUEIDENTIFIER available since PHP 7.0, without it uniqueidentifier is returned as 16 raw bytes
+					return $this->dsn("dblib:charset=utf8;host=$server[host]" . ($port != "" ? ";port=$port" : ($socket != "" ? ";unix_socket=$socket" : "")), $username, $password, $options);
 				}
 			}
 		}